About us refactoring
